Fixed duplicate bit sequence bug and changed from Vector to List
When the device is rotated, the sequences are recreated. However, the old
sequences were not cleared before adding the new ones to the list,
thus causing duplicate bit sequences. This will fix performance. Also, a
Vector was being used to store the sequences. Since synchronization is not
needed, this has been switched to a List for improved performance.